Spring-related renamings following ones in IDEA
